Renamed util.py for unit tests
